# Performance improvements for retessellate