### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
* [x] Does your submission pass tests?

Based on this discussion https://github.com/jscad/OpenJSCAD.org/discussions/1247 I have refactored `retessellate.js` to improve performance.
This function is used by all boolean operations and is the main reason why sometimes they're really slow.
To measure the performance I used a geometry with ~15k polygons which gets reduced to ~7.5k polygons.
Results for current implementation of `retessellate.js` :
- total execution time ~580ms of which:
  -  ~465ms used to determine coplanar polygons (lines 29-39) 
  - ~115ms used to call `reTesselateCoplanarPolygons`

Results for new implementation:
- total execution time ~120ms of which 
  - ~18ms used to determine coplanar polygons (lines 21-22) 
  - ~102ms used to call `reTesselateCoplanarPolygons`

There is an ~5x overall performance improvement, but if we exclude the calls to `reTesselateCoplanarPolygons` there's more like an ~25x improvement.

I haven't added this test to the existing suite because I didn't want it slowing down build times. If you consider it should be added, I can work something out.