@platypii 2023-05-30
> wouldn't this be more appropriate in V3?

Based on the tests being unchanged, it seems like @achirita implemented this in a way which does not change the output result at all. So it is not a breaking API change.

I think anything without breaking changes should go in V2 still. Why? Because V3 is not ready yet, who knows when it will be ready. But users are actively using V2, so improvements are nice, and usage also let's us find and fix bugs sooner.