@z3dev 2023-06-11
@platypii can you follow with lint changes for V2 modeling?